refactor(strategies): lift _events_for_new_steps to _base.py (chunk 1)
Promotes the streaming-event helper from react.py to the shared
strategies module so PlanExecute / ToT / MultiAgent stream()
overrides (next chunks) can reuse it without duplication.
